Fixed data type issue that was causing deserialization error

+++ arangodb-net-standard/ViewApi/Models/ViewConsolidationPolicy.cs
@@ -37,31 +37,31 @@
         /// smaller segments as equal for
         /// consolidation selection (default: 2097152)
         /// </summary>
-        public int? SegmentsBytesFloor { get; set; }
+        public long? SegmentsBytesFloor { get; set; }
 
         /// <summary>
         /// Maximum allowed size of all consolidated
         /// segments in bytes (default: 5368709120)
         /// </summary>
-        public int? SegmentsBytesMax { get; set; }
+        public long? SegmentsBytesMax { get; set; }
 
         /// <summary>
         /// The maximum number of segments that
         /// will be evaluated as candidates 
         /// for consolidation (default: 10)
         /// </summary>
-        public int? SegmentsMax { get; set; }
+        public long? SegmentsMax { get; set; }
 
         /// <summary>
         /// The minimum number of segments that 
         /// will be evaluated as candidates
         /// for consolidation (default: 1)
         /// </summary>
-        public int? SegmentsMin { get; set; }
+        public long? SegmentsMin { get; set; }
 
         /// <summary>
         /// Specified if <see cref="Type"/> is "tier".
         /// </summary>
-        public int? MinScore { get; set; }
+        public long? MinScore { get; set; }
     }
 }
